Rework handling of StreamStats -> Prometheus
Previously Prometheus metrics corresponding to spead2 stream statistics
were only updated when a chunk was received. If the data was all bad and
no chunks were received, they would never get updated.

Instead of pushing counters into prometheus_client, have the registry
call a custom collector to pull counters. Counter updates are still
copied out of the streams into a store in the collector, so that garbage
collection of a stream does not cause those updates to get lost. There
is still a window in which events can go missing (between the last
scrape and garbage collection), but the scraped counter values will
never go backwards.

Closes NGC-531.