Use an asyncio.Queue in xsend
Previously it was using a queue.Queue. If the queue was ever empty it
would deadlock the process, because get_free_heap would block the event
loop waiting for something to be added to the heap, and it would never
be added because only another task on the same event loop would ever add
anything.

I also removed the maximum queue capacity. There is no need to restrict
it, since the finite number of BufferWrappers ensures that the queue has
bounded depth, and it seemed safer not to try to ensure the right
capacity is specified.