Update fgpu descriptor infrastructure
As per comments on PR #199. The sleep intervals between sending
descriptors have been updated, as has the manner in which descriptors
are sent. The descriptors are now also sent once during Engine.start to
ensure all X-Engines are aware of the data, then continually in a
separate descriptors loop.

There are still a few comments to be addressed, which will follow in the
next commit.

Contributes to: NGC-445.