Add baseline test
This one is slightly less sophsiticated than the channelisation one, no
plots output, but it does the job with log messages.

I'm also slightly less than thrilled about jiggling the dsim every time
I step through baselines but it was the easiest way that I know of to
get a new timestamp to wait for.