fgpu: more plumbing for missing data
The calculation now carries through to the OutItem. I did some
refactoring to put all the per-pol data for InItem into a separate
class instead of having a bunch of separate two-element lists.
Unfortunately that did require a bunch of extra asserts to placate numpy
because the vkgdr path means that `samples` and `chunks` are sometimes
None (when the InItem is in the free queue).