Introduce PDF report generation
Also muck about with some of the logic, and include some pytest trickery
not to fail on the first failed assert.

I also update the baseline test to look a bit more formal, including
some verbiage on the docstrings, but I'm not completely satisfied with
how it's turned out.

Relates to NGC-480