Hybrid DDC implementation
N threads cooperatively calculate M results.

This is still not fully optimised because the samples are repeatedly
reloaded from shared memory instead of recycled, but it already shows a
performance improvement.

It still fails on the test with non-power-of-two factors. That's
probably because a number of the divisibility assumptions aren't being
enforced.