Bump coverage and pytest-cov versions in requirements
For some reason, the old versions were causing tests to fail when
running pytest with the --cov flag, but bumping the versions up to the
latest caused things to Just Work (TM). So I didn't look too closely
into why that might be, I just took what I was given and carried on.