Update X-engine's receive transport function
to emulate that of the F-engine more closely. This commit just does the
actual code, no changes made to the unit tests just yet. That's going to
be more of a job.

I also remove the unnecessary add_udp_receiver_transport_functions and
move the logic of adding them to the engine's `start()` and out of the
main script.