Add `free_ringbuffer` argument to `make_streams` for recv
The logic behind this is that the caller knows best how large to make
the allocations on the free ringbuffers (i.e. how many chunks are going
to be added), so it's best to let them just create the ringbuffers and
pass them in as an argument. They're doing that with the data
ringbuffers anyway.

Closes NGC-539.