Create katgpucbf.meerkat module, and put band information there
This limits the amount of copy-paste between the sim_correlator scratch
script, and the new qualification framework.

I am not entirely happy with this because I had wanted katgpucbf itself
not to have to be a dependency for these things, but for all practical
situations it seems that those using these scripts already have
katgpucbf installed. We can undo this later if it becomes a problem.