Consolidate mock_recv_streams fixtures
Along with a clever little trick that Bruce suggested for customising
them for each engine. I quite liked this one..

Also add explicit 127.0.0.1 to the katcp host argument for the xengnie
end-to-end test, as the null string parameter was causing things to
fail. That ought to teach me for making several commits without running
any tests.