Speed up correlate_host
Instead of using a pile of loops with numba, take advantage of highly
optimised matrix multiplication provided by numpy (via BLAS). Even
though this does a lot of extra work (computes the whole correlation
product matrix instead of half, conversions to and from floating point)
it still ends up being twice as fast.

I also tried using scipy.linalg.blas.zherk (which implements A*A^H in
for one half of the output) but it wasn't much faster and involved a lot
of black magic since that's a low-level wrapper around Fortran. It
didn't seem worth it.