Remove references to substream for xengine e2e test
Spead2's documentation suggests that substreams should really be used
when there are multiple output destinations, so using it to "simulate"
multiple input sources doesn't seem quite appropriate.

I did play around with using ant_index as the substream_index as
suggested in the comment, but it ultimately introduced mismatches in the
output between the GPU and the expected. Which I was not in the mood to
investigate. So it seemed better just to remove this.

Also, I removed the arguments passed explicitly to the HeapReference
constructor because they were just the defaults anyway.

Closes NGC-627