Reuse correlators between qualification tests
The sizing fixtures, plus an internal fixture that actually creates a
correlator, are made session-scoped. The `correlator` fixture just takes
the session-scoped fixture and resets the correlator to a known state
(resetting delays and gains, enabling the output stream, and setting the
dsim to output zeros).

This does change the test ordering, such that the correlator config is
the slowest-changing thing, slower than the particular test. Potentially
the reporting should re-sort the data to group together all parametric
versions of a test, but that can be done on the display side.

This depends on #259.

Closes NGC-592.