dsim: use numba to speed up quantisation
This is something like twice as fast.  I think the main benefit is just
from reducing memory traffic (a bunch of steps are done on each element
in turn, rather than slinging temporary arrays around).

This could possibly be extended to doing the bit packing in the same
kernel, which would be even more beneficial for memory bandwidth since
it packs the output down to 10 bit.