dsim: redo how the signal service process is run
Forking seemed to cause some anomalies in the test suite (NGC-637).
Spawning is slower but safer in the presence of features that don't play
nice with forking (e.g. threads and locks).

I reworked the approach to shared memory with a new SharedArray class.
It relies on `os.memfd_create` to create a file handle that can be
passed through the interprocess pipe and mapped on the receiving
process. In retrospect sharedctypes might have worked just fine because
`multiprocessing` has more machinery around it than I realised, but it's
not that much code, and it makes it potenitally possible to use huge
pages in future.

Closes NGC-637.