Rework dither calculation for better performance
There is a lot of refactoring here, but the ultimate goal is that
SignalService computes the dither at startup and reuses it in future.

The Dither subclass of Signal is removed, and replaced with
`make_dither` that directly generates a dask array (wrapped in
xr.DataArray so that the axes can be named). quantise now takes the
dither as input rather than generating it itself.

See NGC-633.