Add infrastructure for antenna_channelised_voltage tests
No tests are added yet, but this provides a utility for measuring the
response to different tones, accelerated by distributing the work across
all the dsims (and both pols of each).

I've tested it with some code that's still a WIP, so I'm reasonably
confident that it's correct, or at least it seems to produce a
reasonable channel shape.