Refactor most fields out of CorrelatorRemoteControl
They have been moved into BaselineCorrelationProductsReceiver, which
will allow different instances to be created for different streams (e.g.
wideband and narrowband).

CorrelatorRemoteControl now uses aiokatcp.SensorWatcher to retrieve all
the sensors, which reduces the need for the hacky get_sensor_val. I also
refactored creation and shutdown of the object to move it into the class
and out of conftest.py (mostly because the SensorWatcher needs to be
added before connecting).

Closes NGC-636.