Update combinations to test in xbgpu
This is mostly to reduce testing time (see NGC-655 for motivation of
choice), but it also adds tests for very few antennas.

The `combinations` pytest marker was semi-rewritten to allow the possible
combinations to be filtered, so that it is possible to test 1 antenna
and to test antenna 3's data being missing, while avoiding testing the
combination of the two (`pytest.skip` would have potentially left some
individual parameter values completely untested).

Closes NGC-655.