Support for shorter periods in dsim
The `?signals` request may optionally specify a period, which must
divide into the number of samples allocated at startup. The generated
signal is then tiled to fill up the memory allocation, which frees the
sending code from worrying about changes in size.

I also pulled out a common unpackbits helper function to unify the 3
different implementations that previously existed, and changed the
--signal-heaps command-line argument to --period (taken in samples)
which is a bit more intuitive.

Closes NGC-659.