Bump versions in pre-commit.yaml
I forgot to do this in the previous commit.

One small wrinkle is that while redis is on 4.3.4, there doesn't seem to
be a release of types-redis for it yet, so that is left at 4.3.3. Since
we're not directly using redis I don't see that being an issue.

Unfortunately mypy is now failing because the newer numpy seems to
incorporate dtypes into its type system, which is causing code that
reassigns arrays with a different dtype to fail.