xbgpu: skip all fully-empty chunks
These phantom chunks are created by spead2. It's going to simplify some
future tests slightly to not have them flow through the pipeline (and
it'll reduce overheads when the correlator starts or after an
interruption). It'll also make warnings about number of missing chunks
more accurate.

One corner case where this makes things worse is that after an
interruption, it will require more chunks to be seen before the
interruption gets logged.