Fix test_delay_model_update_correct unit test
The test was failing due to numeric precision issues caused by using
crazy-large delays (multiple seconds). I think this is also why the test
used to pass despite the code being wrong: the delay was an integer
number of samples (and a multiple of 4), which meant the phase
correction was a multiple of 2pi and thus effectively 0.

Fixed by reducing delays to nanoseconds.