qual: make next_complete_chunk return data only
Instead of returning the chunk, it returns a copy of the data and
immediately returns the chunk to the stream. This simplifies a lot of
code that uses this function.

There will be some performance impact from copying the data, but
generally the tests do not contain performance-critical code immediately
after receiving the chunk.

Closes NGC-703