Use annotations to match requirements to qual tests
Instead of putting requirements as text in the docstring of each test,
use `pytest.mark.requirements` with a list of requirements. The report
generator pulls requirements for requirements_text/{requirement}.tex to
populate the report. In future this will allow us to programmatically
determine which requirements have tests.

The report generation now also maps top-level sections in the test
docstrings to subsubsectios in LaTeX, making them usable. The
"verification method" of each test is now made into a section.