Tidy logic in baseline test
Remove most of the code in the is_signal_expected_in_baseline function,
and fix a logic mistake with `all_expected` (also rename it to
`no_unexpected` for more clarity).

Also, sadly, remove the walrus operators.

This commit also removes the bulk of the actual output of the test,
replacing it with a message stating that no errors have been detected.
Otherwise with an 80-A correlator we would have had about 350 pages of
output, which is perhaps a bit much.