Flatten and split the "Installation and Operation" docs
The structure was

Installation and Operation
-> Installation and Operation
   -> ...

which was silly, and the single .rst was getting unwieldy with 4 levels
of headings. I've split it into separate documents, and also pulled
Monitoring up a level (out from under Control) to flatten things
further. There are no content changes other than switching the
underlining used on headings to reflect the shallower structure.