Describe general DSP engine operation
I generalised most of the generic stuff in the fgpu.design document, and
moved it to `engines`. Then I hacked the corresponding generic stuff out
of xbgpu.design. The remainder stayed in fgpu.design. xbgpu.design got
rearranged a bit, I moved the description of the accumulation and output
to after the description of the tensor core implementation details, it
seemed to make most logical sense for that.

The images were mostly not needed, so I deleted them. This probably
removes the need for NGC-524 for the most part.

I also noticed that missing data handling wasn't described for xbgpu, so
I added a TODO for that to be added, and an accompanying ticket.

Closes NGC-675. I think.