Replace DeclareMathOperator with operatorname in maths.rst
This works fine when rendering to HTML, but when rendering to latexpdf
it does not, becuase the `\DeclareMathOperator` directive is expected
only in the preamble of the document.

Moving it to `latex_elements["preamble"]` fixed the latexpdf output, but
broke HTML because mathjax doesn't have access to that directive
anymore.

So this commit simply uses the `\operatorname` directive instead, which
is perhaps somewhat less elegant, but works for both outputs.