Redesign delay model
Instead of interpreting the start time as an input timestamp, interpret
it as an output timestamp. Similarly, interpret the delay rate as
advancing a certain amount per output tick, instead of per input tick.
This removes the need to "invert" the delay model; it simply gives you
delays as a function of output time, instead of input time.

This greatly simplifies a lot of the code, but there are probably some
lingering bugs when a MultiDelayModel is not monotonic. Either it needs
to be adapted to make it so (by tweaking LinearDelayModel to take a
minimum input timestamp), or engine.py needs to be robust to this
occurring. It might actually be fine, since a backwards jump implies a
change in coarse delay, which already causes it to interrupt a batch of
work.