Partial implementation of saturation calculations
The bulk of this commit is actually refactoring to avoid needing `main`
to itself sample the initial signal before it constructs the
DeviceServer. That was previously necessary because the sender needed to
know the initial timestamp, which can only be determined after the
initial signal is computed (because it is slow). This dependency is
broken by having Sender take the initial timestamp and sync time as
arguments to `run` rather than in the constructor.

The sampling functions now (optionally) compute a set of saturation
flags. These still need to be reduced to produce saturation counts and
transmitted. This functionality also requires testing.