Create xsend.Heap class
It takes the previous tuple of a buffer and a future and bundles it with
a few more things into a class. It now contains the saturation count and
puts it into a Prometheus counter on send (still untested), and the heap
is constructed in advance (cheaper, although I don't expect it to make
any practical difference).

I also removed a lot of unused fields from XSend.