Rework DeviceStatusSensor with SimpleAggregateSensor
This simplifies the logic somewhat, and makes the timestamp logic the
responsibility of aiokatcp. A functional change is that now if the worst
status is beyond ERROR (e.g. UNREACHABLE), the aggregate status will be
ERROR. That shouldn't have any effect since katgpucbf does not provide
any such sensors.

This will fail to run right now because it needs a new release of
aiokatcp that implements SimpleAggregateSensor.