Bump PyCUDA to 2022.2
This allows the workaround for the build-time dependency on numpy to be
dropped, as 2022.2 uses oldest-supported-numpy instead.

Also update katsdpsigproc to 1.6.1 which contains a fix for a bug that
was only exposed when using the newer pycuda. I don't know whether the
newer pycuda is only showing a new warning (and the old one silently
leaked) or if the old version didn't provoke the leak at all.