Update xbgpu recv_chunks for new Rx sensors
Specifically in the function signature in what it requires, and all its
uses in the package - the receiver loop and unit tests.

Still need to update the unit tests themselves for checking of the
sensors after data moves through the pipeline. But the plumbing seems to
be hooked up correctly.

Contributes to: NGC-805.