Remove most content from requirements.in files
Instead, use pip-compile's ability to gather requirements from
setup.cfg.

The remaining stuff are things that are not strictly needed to run the
package, but that are either generally useful (e.g. ipython in
requirements-dev) or are wanted in a deployed scenario (psutils).

This increases the size of requirements-dev.txt, because it now
includes all the non-optional requirements of katgpucbf. That should
make any difference when one installs both requirements.txt and
requirements-dev.txt, since they'll just be duplicated.

qualification/requirements.txt gets a lot smaller, because it no longer
pulls in all the documentation stuff that was in requirements-dev.txt
(previously it just included requirements-dev.txt with `-r`).