Amortise some unpack calculations
A lot of the work in figuring out how to unpack in the PFB kernel is
fixed for each thread (because the step size is a multiple of 8). So it
can be precomputed and stored in an unpack_t struct.

This is a pretty minor optimisation since the kernel is not
compute-bound, but it does improve things by about 1%. It may also
facilitate some future optimisations.