Optimise the PFB FIR
The "clever" approach to writing samples into a circular buffer using
modular arithmetic (and relying on the compiler to collapse it unrolling
the loop) seems to perform worse than just shunting all the samples one
along every iteration. I suspect in the latter case the compiler
eliminates a lot of the copies anyway. It also seems performance is
improved by letting the compiler decide on the loop unrolling instead of
forcing a specific factor.

Also speed up the dot product fractionally by initialising it with the
first product, rather than with zero.

Fully spell out a couple of types just for consistency.

This leads to an 8% speedup in some cases.