Modify preamble and katdoc.sty with latexindent
No functional changes (yet), just the indentation of the code.

I'm busy working on adding some functionality but I'm going to leave
that into a subsequent commit to separate it by itself.

I've been working on a more streamlined version of katdoc.sty elsewhere
that I'm probably going to merge into here as well, this is just an
interim step.