Rename doc to report_doc, append _report to pdf name
The idea behind this is, pretty soon we're going to be generating two
docs, one report and one procedure. This is just a light refactor
to prepare for that.

The convention that we've had of just naming everything "report.pdf" is
now going to be slightly weird, it'd come out `report_report.pdf`. I
have changed Jenkins to use `test` in future, (the `.pdf` is just
stripped anyway) in which case we'll get `test_report.pdf` which will
later to be accompanied by `test_procedure.pdf`.