Rename procedure section to test_detail section
In the detailed test results, there was a chunk called "procedure" which
basically just reports on the procedure that was followed.

This will cause ambiguity in the future, since I'm going to be referring
to a "procedure" and a "report" as two separate documents, and building
them up in parallel.

This isn't quite a refactor as I do actually change the heading in the
output document, but it's fairly minor.

Also renamed a context manager named `section` to a slightly
less-ambiguous `detailed_results`, based on its actual text heading.

I also fiddled with a context manager that looked a little out of place.
The way PyLatex works doesn't seem to have made a difference but it made
me feel better about the structure of the code.