Add parallel generation of test procedure document
Currently this is my first attempt at this.

It involved a bit of a rearrangement of stuff to group together things
that are common to both the report and the procedure doc, and then
instead of doing a test detail with `step` and `detail`/`figure`
whatever, I just append the `step`s directly.

I did this just to make sure that they all actually get there, to avoid
brain frying for now. This works, so I figured, let's get it committed.

I think that probably the most elegant thing to output would be an
enumerated list because there are nice numbers, but I'll handle that
challenge next.