Support separated input and output network interfaces
On input, an interface can be specified per pol. On output, an arbitrary
number of interfaces may be specified, and the multicast groups will be
distributed between them (although descriptors and stop messages will be
sent on the first interface).

This will still need some unit tests, particularly because there is a
fast path in place to minimise risk to existing performance.