Use multiple command queues
Run the PFB FIR and the FFTs in two separate command queues. This allows
for greater utilisation of the GPU, particularly when the chunk size is
small to accommodate vkgdr.

It won't work out of the box with release katsdpsigproc, because
katsdpsigproc currently prevents an OperationSequence from mixing
operations with different command queues.