Eliminate a round trip through the CPU
We need to wait for the PFB FIR to run on an InItem's memory before
starting the next transfer to it. That was previously done by waiting on
the CPU, but it can be handled by an inter-queue dependency so that the
Python code doesn't need to be woken up for it.