Make raw data in report.json optional
It takes a lot of space, so gate it with an ini option (raw_data).

Also fix a bug that meant use_ibv was treated as true if omitted,
because it defaults to a string "false" which is truthy. And fix another
bug that prevented setting any setting to a falsey value.