Use cross-correlations to better measure channel shape
The channel shape was being limited by the dither/quantisation noise
in the dsim. Because auto-correlations were being used, this noise was
self-correlated. Using cross-correlations, with the same tone but
independent noise, means the noise is uncorrelated and so tends to wash
out.

The implementation uses the two pols of each dsim as the two sources to
correlate. This simplifies the code somewhat, but halves the efficiency
as it now requires two inputs to do what previous was done with one. To
compensate, I've bumped DEFAULT_ANTENNAS up to 8.