Rearrange report generation in qualification Jenkinsfile
There was a try/finally block. The only thing that was in 'try' was the
pytest run itself. But if there's a problem with the report generation,
this doesn't get caught.

So I moved the execution of script to generate the report (and in fact
the procedure document) into the `try` block, added a catch just to
publish the json if everything else goes wrong, and then the finally
remains basically as it was for whatever is left.

I have no idea if this will work, but I'm unable to retrieve the JSON
from failed Jenkins runs so this is my attempt at debugging that.

Contributes to NGC-875