Nest try/catch/finally block in Jenkinsfile
I wasn't quite happy with the way that I had it in the previous commit.
Sure, the intermediate JSON was being published. But the way that I was
catching the exception let Jenkins think that there wasn't actually a
test failure, so it was coming back as "good".

On the plus side, the JSON did get published, but it's large, some 200
MB in size. So in order to reduce that, this commit asks Jenkins to
compress it using `xz` which is installed on qgpu02, it may be a fairly
standard Linux thing. `zstd` wasn't, so I went with `xz`. In my test, it
reduces it from 200MB to about 8. This closes NGC-895 if it works.