Force compression
Jenkins has been complaining on the output of this command that the file
already exists.

This suggests that it's not actually clearing up its workspace after
previous runs. But I'm not sure where it is, so I can't actually check.

This should at least make sure that you always get the latest one.