Fix test_delay_application_time
The test sets a delay model for 200ms in the future, then tries to grab the
dump that brackets that change. The *start* of that dump may be in the
past, so (implicitly) using steady_state_timestamp means we might just
throw away the dump we actually want.

I'm not sure why we've only run into this with UHF and not with other
bands.

Partly addresses NGC-897.