Add --wideband command-line argument
This replaces the previous --channels, --taps, --w-cutoff and dst args
which configured a single wideband stream. This is thus an incompatible
change which will need to be coordinated with a change to
katsdpcontroller.

It also highlighted a bug in the previous multi-stream PR where
chunk_jones was only taking the singleton wideband stream into account
for rounding purposes.