fgpu: make some katcp sensors pipeline-specific
Prefix the names of eq, delay, feng-clip-cnt and dig-pwr-dbfs sensors
with the output name. The last of these is done for expediency: it ought
to be roughly the same for all outputs, but it's calculated
per-pipeline, and probably not on the identical set of inputs and at the
identical cadence due to the interaction with PFB windows.

See NGC-569.