Load narrowband filter coefficients
The coefficients used for 1/8 and 1/16 modes are the ones used in
MeerKAT. For other ratios the user is required to specify the filter
design parameters with new command-line arguments (w_pass, w_stop,
weight_pass).

Additionally, the number of DDC filter taps can now be overridden from
the command line (ddc_taps).