Support complex PFB in postprocessing kernel
Apart from the complex_pfb support, the processing of negative
frequencies are also rethought a bit. Previously we computed the
conjugate of negative frequencies because that was slightly more
convenient for fix_r2c. But for complex_pfb we would have had to undo
that, and emitting them properly actually simplified the code a little.

This does not yet correctly handle delays. In particular, the phase will
be referenced to the centre of the narrowband, making the phase-juggling
arithmetic moot. It could be further corrected here, or the rest of the
code could be adjusted to match this, which would allow for some
simplifications elsewhere.

Fine delays are probably also in the wrong units (original versus
subsampled samples).

Closes NGC-923.