Change how phase is aligned to centre frequency
Previously, the delay model was adjusted on input so that the phase was
referenced to channel 0 (and this was undone on query). Instead, keep
the delay model as is, and update the postproc kernel to apply the delay
slope through the centre frequency instead of channel 0. This still
needs a correction term to undo the phase effect of coarse delay.

The main motivation is that it makes narrowband simpler, where the mixer
already shifts the centre frequency to DC, and thus it is more natural
to use the original delay model. The correction that is now introduced
for wideband would have been needed for narrowband anyway.

Also fix the scaling of fine_delay for narrowband.