Use twice as much internal bandwidth for narrowband
Produce double the number of channels internally, then discard half of
them. This allows for a much better channel response.

This is an early implementation which produces the doubled output all
the way to the end of GPU processing. Future updates will modify the
postproc kernel to cut out the unwanted data earlier.

See NGC-939.