Move discard-half into postproc kernel
The postproc kernel now emits only the narrowband channels that will be
transmitted. The main benefit is likely to be GPU memory usage (OutItem
is half as large), although it may be a little faster due to less
bandwidth usage. It will also now get saturation counts correct.

This could probably be optimised quite a lot, since the kernel still
does almost all the work for the channels that will be discarded.