Bump almost all requirements to latest
This was prompted by Dependabot moaning about a security issue on one
package, but I figured it was about time to do a general upgrade.

I kept back numba/llvmlite because I was getting warnings about the TBB
version on Ubuntu 22.04 being too old. That in turn forced numpy to be
kept back.

A few other changes:
- Removed types-setuptools from pre-commit. It was there for
  pkg_resources, which has now been eliminated.
- Removed `--extra=vkgdr` from pre-commit of requirements.txt, because
  vkgdr is now a hard requirement.

The new version of black made a couple of code formatting changes.