Redo new DDC kernel
The decoder struct is replaced by a decoder function. This makes the
retained state explicitly just a (per-unroll) buffer and a (per-thread)
load offset. Everything else is determined from compile-time constants.

This does introduce the requirement that C*SUBSAMPLING*SAMPLE_BITS must
be a multiple of SAMPLE_WORD_BITS, but since C=8 seems to work well,
DIG_SAMPLE_BITS is even and SUBSAMPLING needs to be even for other parts
of the engine to work, this shouldn't be a major problem.