Make PFB kernel load samples as words, not bytes
This doesn't noticeably affect the speed, but it will make it easier to
partially merge the code with that in the DDC kernel, which also uses
words. It also allows any number of sample bits to be used (up to 32),
although DIG_SAMPLE_BITS_VALID is limited to 16 as larger values can
overflow the total_power counter.