Run wideband+narrowband in test_engine
Instead of only running the desired output stream (which does not test
the real-world use of running both concurrently), run both and only do
tests against one at a time.

Ththe main functional change in the test is that previously
--dst-chunk-jones was set differently for wide and narrow, but now they
need to have a single value. I kept the value picked for narrowband just
because that made all the tests pass.

Closes NGC-955.