fgpu engine tests: use non-round nb centre frequency
Test with a narrowband centre frequency which is not a multiple of the
channel width. This requires adding extra phase calculations to the
tests because the tested tone frequencies are also not multiples of the
channel width, and hence do not perfectly wrap over
samples_between_spectra samples.

Also added a helper function to compare arrays of angles.