Use fixed-point arithmic for DDC mixer phase
This eliminates some floating-point consistency issues, where results
would change slightly if the tuning parameters (specifically unrolling)
changed. This approach is also a few percent cheaper as it does not
require double-precision arithmetic in the kernel.

The downside is that the requested centre frequency is quantised
slightly, and this will require changed to katsdpcontroller to report
the actual frequency used. The quantisation is to a multiple of
output-bw/2**32, which is far below the channel width.

Closes NGC-991.