Update ingest script
It was written quite a while ago and relied on some assumptions that are
no longer true. First among these is outdated sensor names, but also
some logic from the derived classes of Chunks in katgpucbf seems to have
found its way here. Which isn't how the base spead2 Chunks worked.

I'm not quite sure how it got to that state, but it's fixed now. I was
able to ingest and plot spectra.

Closes NGC-1007