Increase Jenkinsfile timeout for `pytest quick`
Probably due to the number of unit tests we've added lately, we started
seeing timeouts with the `pytest quick` stage of the CI run.

In ideal settings, when the GPU isn't doing anything else, the tests
take about 6 and a half minutes. This commit bumps up to 15, to
accommodate two simultaneous builds that occupy the GPU, plus a little
bit of headroom for in case.

Closes NGC-1009.