Add pre-tuning support
A pre-tuning database is added to the repository, and the Dockerfile
inserts it into the image so that it will be used at runtime. This
avoids every container re-doing autotuning (provided the parameters are
in the space that's pre-tuned), which could be problematic when run by
multiple containers simultaneously (skewing tuning results).

Re-tuning currently requires
https://github.com/ska-sa/katsdpsigproc/pull/108, which is merged but
not yet released and hence not yet present in the Docker image. The
tuning results committed so far used a workaround to mount that version
into the container.

Closes NGC-983.