Remove CHANNEL_OFFSET in favour of a frequency fixture
Previously, we'd been running the X-engine end-to-end test as
(arbitrarily) the 5th X-engine (i.e. with index 4). Which is fine,
except when we get to small array sizes when there are far fewer
X-engines than that. So I adjusted this to a dynamic-ish fixture that
scales according to the number of X-engines, so that the test will
always be roughly in the middle of the channel range.

I came across this issue while looking for another issue (NGC-929
specifically). This wasn't the problem, but I thought it warranted
fixing.