xbgpu: Call QueueItem.reset when item is acquired, not released
This gives an opportunity to await any events that came with the item on
the free item queue. It fixes a race condition introduced as part of the
work on NGC-929, where events were getting lost when `reset()` cleared
the event list (there is a bigger issue that this is a foot-gun, but
that's a problem for another day).

I also changed tx_items to work the same way, although it should be a
non-issue since all work on the TxItem is synchronously completed before
it goes onto the free queue.

Tests are still failing with an out of resources error, but I think
that's unrelated.