Clamp dig-rms-dbfs sensor to DBL_MIN in case of zero power
Or rather strictly speaking it's to np.finfo(np.float64).min, which I
think may be the same as DBL_MIN under the hood, though it's probably
implementation dependent.

The context to this is that previously in the case of zero input power
we'd assign a -inf value. This is OK for the IEEE double specification,
but the KATCP spec only allows decimal values for float sensors, so we
weren't compliant in this sense. Issues arose with CAM's NATS system
which didn't like the -inf values.

I couldn't find any suitable unit test where these sensors were
exercised, so I added a simple one.

Closes NGC-1059